H1: send confirmation email: basics
Unlike the book, we had to manually delete the file

  app/views/order_notifier/received.html.erb

because its existence made Rails automatically choose it over the
`received.text.erb` file we created to send plaintext emails. This
`received.html.erb` file was auto-generated by the `rails generate
mailer` command.

Also, although the mailing mechanism has been tested, we have not
commited the username and password of the smtp settings in

  config/environments/development.rb

for security reasons.