order_notifier_test: fix fixtures for mailers test
The changes to the test fixtures are needed to make the mailer tests
pass as-is under test/mailers/order_notifier_test.rb; previously, the
assert_match regexes would fail because of blank products (line items).
Now, the line items are populated as expected.

Historically, the file `line_items.yml` was created as part of the
`rails generate scaffold LineItem product:references cart:belong_to`
command back in bbe46dc7fd28a00da220ea0a3db9b52c216c89b4.

In 5d8103ca88efd9192807c9db9991127cf40b48e5, we added the values to this
file (the value fields after the colons were blank).

It appears that in "Iteration G1: Capturing an Order" in the book (p.
167) there is a link to a "fixed" `line_items.yml`. Obviously, we had
overlooked this file. Interestingly, however, that section is the book
is the only time that `line_items.yml` is referenced as a source link,
and its contents are thus:

    # Read about fixtures at
    # http://api.rubyonrails.org/classes/ActiveRecord/Fixtures.html

    one:
      product: ruby
    #START_HIGHLIGHT
      order: one
    #END_HIGHLIGHT

    two:
      product: ruby
      cart: one

But to make the mailer tests pass, we only had to change some of the
lines in `line_items.yml`. The "<model>_id" syntax was probably chosen
by a more recent version of Rails, rather than the "<model>" syntax used
in the book (e.g., "product_id" instead of "product").